Centralize the save PR logic.
